fix: use float16 instead of bfloat16 in FluxProvider

+++ backend/ai_providers/flux_provider.py
@@ -43,7 +43,7 @@ class FluxProvider(BaseAIProvider):
             logger.info("Caricamento pipeline Flux...")
             model_path = self.model_info.get("path")
             try:
-                self.pipeline = FluxPipeline.from_pretrained(model_path, torch_dtype=torch.bfloat16)
+                self.pipeline = FluxPipeline.from_pretrained(model_path, torch_dtype=torch.float16)
                 self.pipeline.enable_vae_tiling()
                 self.pipeline.enable_vae_slicing()
                 self.pipeline.enable_attention_slicing()
@@ -63,7 +63,7 @@ class FluxProvider(BaseAIProvider):
                     raise RuntimeError(f"RAM di sistema insufficiente ({available_ram:.2f}GB) per il fallback su CPU. Flux richiede circa 24GB di RAM. Operazione annullata per evitare il blocco del sistema.")
                 
                 logger.warning(f"RAM disponibile: {available_ram:.2f}GB. Uso model CPU offload per evitare OOM.")
-                self.pipeline = FluxPipeline.from_pretrained(model_path, torch_dtype=torch.bfloat16)
+                self.pipeline = FluxPipeline.from_pretrained(model_path, torch_dtype=torch.float16)
                 self.pipeline.enable_vae_tiling()
                 self.pipeline.enable_vae_slicing()
                 self.pipeline.enable_model_cpu_offload(gpu_id=gpu['device_index'])
